requirements.txt
geocoder==1.38.1
Pillow==9.4.0
Pillow==10.0.0
Requests==2.31.0
